+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Linstor/Drbd/OutsideAddress property for DRBD's outside-address for setups using a NAT
 - Added optional alternative suffixes for external files
 - Added status check (current path + content as expected) for external files
+- Added property "Rest/AutoHTTPs", if enabled will automatically create a SSL certificate and re-start SSL server
 
 ### Changed